refactored qunit tests and got them all to pass

+++ app/assets/javascripts/controllers/course/course_index_controller.js
@@ -1,19 +1,26 @@
 App.CourseIndexController = Ember.ObjectController.extend(Ember.Validations.Mixin, {
 
-	isOwner: function() {
+	userIsOwner: function() {
 		var user_id = this.get('model').get('user_id');
 		return (App.AuthManager.isAuthenticated() && (App.AuthManager.get('apiKey.user.id') == user_id));
 	},
 
 	actions: {
 		delete: function(course) {
-			var userConfirm = confirm("Are you sure you want to delete the course '" + course.get('name') + "'?");
-			if(userConfirm){
-				course.destroyRecord();
-				this.transitionToRoute('courses');
-			} else {
-				this.transitionToRoute('course', course);
+			if(this.get('userIsOwner')) {
+				var userConfirm = confirm("Are you sure you want to delete the course '" + course.get('name') + "'?");
+				if(userConfirm) {
+					course.destroyRecord();
+					this.transitionToRoute('courses');
+				} 
+				else {
+					this.transitionToRoute('course', course);
+				}
 			}
+			else {
+				alert("You must own the course to delete it.");
+				this.transitionToRoute('course', course);
+			}	
 		},
 
 		edit: function(course) {
